Move linear operators into LinearOperators
