+++ src/NumericalAlgorithms/CMakeLists.txt
@@ -1,5 +1,6 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
+add_subdirectory(LinearOperators)
 add_subdirectory(RootFinding)
 add_subdirectory(Spectral)
