+++ src/NumericalAlgorithms/LinearOperators/CMakeLists.txt
@@ -1,10 +1,12 @@
 # Distributed under the MIT License.
 # See LICENSE.txt for details.
 
-set(LIBRARY Spectral)
+set(LIBRARY LinearOperators)
 
 set(LIBRARY_SOURCES
-    LegendreGaussLobatto.cpp
+    DefiniteIntegral.cpp
+    Linearize.cpp
+    MeanValue.cpp
     )
 
 add_library(${LIBRARY} ${LIBRARY_SOURCES})
@@ -13,4 +15,5 @@ target_link_libraries(
   ${LIBRARY}
   INTERFACE DataStructures
   INTERFACE ErrorHandling
+  INTERFACE Spectral
   )
