+++ src/NumericalAlgorithms/LinearOperators/Linearize.cpp
@@ -1,7 +1,7 @@
 // Distributed under the MIT License.
 // See LICENSE.txt for details.
 
-#include "NumericalAlgorithms/Spectral/Linearize.hpp"
+#include "NumericalAlgorithms/LinearOperators/Linearize.hpp"
 
 #include <array>
