+++ tests/Unit/Numerical/Spectral/CMakeLists.txt
@@ -2,12 +2,9 @@
 # See LICENSE.txt for details.
 
 set(SPECTRE_UNIT_SPECTRAL_TESTS
-    Numerical/Spectral/Test_DefiniteIntegral.cpp
-    Numerical/Spectral/Test_LegendreGaussLobatto.cpp
-    Numerical/Spectral/Test_Linearize.cpp
-    Numerical/Spectral/Test_MeanValue.cpp
-    )
+  Numerical/Spectral/Test_LegendreGaussLobatto.cpp
+  )
 
 set(SPECTRE_UNIT_NUMERICAL_TESTS
-    "${SPECTRE_UNIT_NUMERICAL_TESTS};${SPECTRE_UNIT_SPECTRAL_TESTS}"
-    PARENT_SCOPE)
+  "${SPECTRE_UNIT_NUMERICAL_TESTS};${SPECTRE_UNIT_SPECTRAL_TESTS}"
+  PARENT_SCOPE)
